+++ src/main/kotlin/com/conrradocamacho/alugames/util/GameExtension.kt
@@ -4,5 +4,5 @@ import com.conrradocamacho.alugames.model.Game
 import com.conrradocamacho.alugames.model.InfoGameJson
 
 fun InfoGameJson.createGame(): Game {
-    return Game(this.titulo, this.capa, this.preco, this.descricao)
+    return Game(this.titulo, this.capa, this.preco.toBigDecimal(), this.descricao)
 }
